+++ tests/bigquery/data/sample_log.json
