--- tests/bigquery/lineage/data/view_result.json
@@ -1,38 +1,0 @@
-[
-  {
-    "logicalId": {
-      "name": "project1.dataset1.table1",
-      "platform": "BIGQUERY"
-    },
-    "entityUpstream": {
-      "sourceEntities": [
-        "DATASET~A6BA6F986B360A57CF65200F29F5B251"
-      ],
-      "transformation": "select * from `foo`"
-    }
-  },
-  {
-    "logicalId": {
-      "name": "project1.dataset1.table2",
-      "platform": "BIGQUERY"
-    },
-    "entityUpstream": {
-      "sourceEntities": [
-        "DATASET~A6BA6F986B360A57CF65200F29F5B251"
-      ],
-      "transformation": "select * from `Foo`"
-    }
-  },
-  {
-    "logicalId": {
-      "name": "project1.dataset1.table3",
-      "platform": "BIGQUERY"
-    },
-    "entityUpstream": {
-      "sourceEntities": [
-        "DATASET~A6BA6F986B360A57CF65200F29F5B251"
-      ],
-      "transformation": "select * from foo"
-    }
-  }
-]
